backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
email-validator==2.1.0.post1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis==5.0.1
psycopg2-binary==2.9.9
httpx==0.26.0
pytz==2024.1
celery==5.3.6
requests==2.31.0
stix2==3.0.1
weasyprint==61.2
Jinja2==3.1.3
python-telegram-bot==20.8
